Added: Added game setting to enable verbose logging
Addes a new setting for verbose logging in the game settings.  Defaults to FALSE.

Wrapped calls to ROLLog.log and ROLLog.debug up in a check against the user's verbose log setting and only do the logging if set.

Logging has a performance hit, and even preparing a string to log can cause boxing and allocations.